+++ secator/installer.py
@@ -24,7 +24,7 @@ from secator.definitions import OPT_NOT_SUPPORTED
 from secator.output_types import Info, Warning, Error
 from secator.rich import console
 from secator.runners import Command
-from secator.utils import get_versions_from_string
+from secator.utils import debug, get_versions_from_string
 
 
 class InstallerStatus(Enum):
@@ -66,7 +66,7 @@ class ToolInstaller:
 		# Fail if not supported
 		if not any(_ for _ in [
 			tool_cls.install_pre,
-			tool_cls.install_github_handle,
+			tool_cls.github_handle,
 			tool_cls.install_cmd,
 			tool_cls.install_post]):
 			return InstallerStatus.INSTALL_NOT_SUPPORTED
@@ -86,12 +86,23 @@ class ToolInstaller:
 
 		# Install binaries from GH
 		gh_status = InstallerStatus.UNKNOWN
-		if tool_cls.install_github_handle and not CONFIG.security.force_source_install:
-			gh_status = GithubInstaller.install(tool_cls.install_github_handle, version=tool_cls.install_version or 'latest')
+		install_ignore_bin = get_distro_config().name in tool_cls.install_ignore_bin
+		if tool_cls.github_handle and tool_cls.install_github_bin and not CONFIG.security.force_source_install and not install_ignore_bin:  # noqa: E501
+			gh_status = GithubInstaller.install(
+				tool_cls.github_handle,
+				version=tool_cls.install_version or 'latest',
+				version_prefix=tool_cls.install_github_version_prefix
+			)
 			status = gh_status
 
 		# Install from source
 		if not gh_status.is_ok():
+			# Install pre-required packages
+			if tool_cls.install_cmd_pre:
+				status = PackageInstaller.install(tool_cls.install_cmd_pre)
+				if not status.is_ok():
+					cls.print_status(status, name)
+					return status
 			if not tool_cls.install_cmd:
 				status = InstallerStatus.INSTALL_SKIPPED_OK
 			else:
@@ -202,18 +213,36 @@ class SourceInstaller:
 
 		# Install build dependencies if needed
 		if install_prereqs:
-			if 'go ' in install_cmd:
-				status = PackageInstaller.install({'apt': ['golang-go'], '*': ['go']})
-				if not status.is_ok():
-					return status
-			if 'gem ' in install_cmd:
-				status = PackageInstaller.install({'apk': ['ruby', 'ruby-dev'], 'pacman': ['ruby', 'rubygems'], 'apt': ['ruby-full', 'rubygems']})  # noqa: E501
-				if not status.is_ok():
-					return status
-			if 'git ' in install_cmd or 'git+' in install_cmd:
-				status = PackageInstaller.install({'*': ['git']})
-				if not status.is_ok():
-					return status
+			regex = re.compile(r'(cargo\s+|go\s+|gem\s+|git\s+)')
+			matches = regex.findall(install_cmd)
+			matches = list(set(matches))
+			for match in matches:
+				match = match.strip()
+				if match == 'cargo':
+					status = PackageInstaller.install({'*': ['curl']})
+					if not status.is_ok():
+						return status
+					rust_install_cmd = 'curl https://sh.rustup.rs -sSf | sh -s -- -y'
+					distribution = get_distro_config()
+					if not distribution.pm_installer:
+						return InstallerStatus.UNKNOWN_DISTRIBUTION
+					if distribution.pm_name == 'apk':
+						install_cmd = install_cmd.replace('cargo ', 'RUSTFLAGS="-Ctarget-feature=-crt-static" cargo ')
+					status = SourceInstaller.install(rust_install_cmd)
+					if not status.is_ok():
+						return status
+				if match == 'go':
+					status = PackageInstaller.install({'apt': ['golang-go'], '*': ['go']})
+					if not status.is_ok():
+						return status
+				if match == 'gem':
+					status = PackageInstaller.install({'apk': ['ruby', 'ruby-dev'], 'pacman': ['ruby', 'rubygems'], 'apt': ['ruby-full', 'rubygems']})  # noqa: E501
+					if not status.is_ok():
+						return status
+				if match == 'git':
+					status = PackageInstaller.install({'*': ['git']})
+					if not status.is_ok():
+						return status
 
 		# Handle version
 		if '[install_version]' in install_cmd:
@@ -232,7 +261,7 @@ class GithubInstaller:
 	"""Install a tool from GitHub releases."""
 
 	@classmethod
-	def install(cls, github_handle, version='latest'):
+	def install(cls, github_handle, version='latest', version_prefix=''):
 		"""Find and install a release from a GitHub handle {user}/{repo}.
 
 		Args:
@@ -242,8 +271,9 @@ class GithubInstaller:
 			InstallerStatus: status.
 		"""
 		_, repo = tuple(github_handle.split('/'))
-		release = cls.get_release(github_handle, version=version)
+		release = cls.get_release(github_handle, version=version, version_prefix=version_prefix)
 		if not release:
+			console.print(Warning(message=f'Could not find release {version} for {github_handle}.'))
 			return InstallerStatus.GITHUB_RELEASE_NOT_FOUND
 
 		# Find the right asset to download
@@ -258,7 +288,7 @@ class GithubInstaller:
 		return cls._download_and_unpack(download_url, CONFIG.dirs.bin, repo)
 
 	@classmethod
-	def get_release(cls, github_handle, version='latest'):
+	def get_release(cls, github_handle, version='latest', version_prefix=''):
 		"""Get release from GitHub.
 
 		Args:
@@ -273,11 +303,12 @@ class GithubInstaller:
 		if version == 'latest':
 			url = f"https://api.github.com/repos/{owner}/{repo}/releases/latest"
 		else:
-			url = f"https://api.github.com/repos/{owner}/{repo}/releases/tags/{version}"
+			url = f"https://api.github.com/repos/{owner}/{repo}/releases/tags/{version_prefix}{version}"
 		headers = {}
 		if CONFIG.cli.github_token:
 			headers['Authorization'] = f'Bearer {CONFIG.cli.github_token}'
 		try:
+			debug(f'Fetching release {version} from {url}', sub='installer')
 			response = requests.get(url, headers=headers, timeout=5)
 			response.raise_for_status()
 			latest_release = response.json()
@@ -289,8 +320,8 @@ class GithubInstaller:
 			return None
 
 	@classmethod
-	def get_latest_version(cls, github_handle):
-		latest_release = cls.get_release(github_handle, version='latest')
+	def get_latest_version(cls, github_handle, version_prefix=None):
+		latest_release = cls.get_release(github_handle, version='latest', version_prefix=version_prefix)
 		if not latest_release:
 			return None
 		return latest_release['tag_name'].lstrip('v')
@@ -441,13 +472,14 @@ def parse_version(ver):
 		return None
 
 
-def get_version_info(name, version_flag=None, install_github_handle=None, install_cmd=None, install_version=None, version=None, bleeding=False):  # noqa: E501
+def get_version_info(name, version_flag=None, github_handle=None, install_github_version_prefix=None, install_cmd=None, install_version=None, version=None, bleeding=False):  # noqa: E501
 	"""Get version info for a command.
 
 	Args:
 		name (str): Command name.
 		version_flag (str): Version flag.
-		install_github_handle (str): Github handle.
+		github_handle (str): Github handle.
+		install_github_version_prefix (str): Github version prefix.
 		install_cmd (str): Install command.
 		install_version (str): Install version.
 		version (str): Existing version.
@@ -488,12 +520,16 @@ def get_version_info(name, version_flag=None, install_github_handle=None, instal
 		info['latest_version'] = str(ver)
 		info['install_version'] = str(ver)
 		info['source'] = 'supported'
-		latest_version = str(ver)
+		if ver:
+			latest_version = str(ver)
 	else:
 		latest_version = None
 		if not CONFIG.offline_mode:
-			if install_github_handle:
-				latest_version = GithubInstaller.get_latest_version(install_github_handle)
+			if github_handle:
+				latest_version = GithubInstaller.get_latest_version(
+					github_handle,
+					version_prefix=install_github_version_prefix,
+				)
 				info['latest_version'] = latest_version
 				info['source'] = 'github'
 			elif install_cmd and install_cmd.startswith('pip'):
@@ -508,6 +544,7 @@ def get_version_info(name, version_flag=None, install_github_handle=None, instal
 							version = max(version, ver)
 							latest_version = str(version)
 							info['source'] = 'pypi'
+				version = str(version) if version else None
 			else:
 				info['errors'].append('Cannot get latest version for query method (github, pip) is available')
 	info['latest_version'] = f'v{latest_version}' if install_version and install_version.startswith('v') else latest_version  # noqa: E501