+++ secator/utils_test.py
@@ -70,6 +70,9 @@ INPUTS_TASKS = {
 	PATH: '.',
 	DOCKER_IMAGE: 'redis:latest',
 	GIT_REPOSITORY: 'https://github.com/freelabz/secator',
+	'gf': 'http://testphp.vulnweb.com/hpp?pp=1',
+	'maigret': 'Linus__Torvalds',
+	'searchsploit': 'apache',
 }
 
 #---------------------#
@@ -119,6 +122,7 @@ META_OPTS = {
 	'trivy.output_path': load_fixture('trivy_output', FIXTURES_DIR, only_path=True),
 	'wafw00f.output_path': load_fixture('wafw00f_output', FIXTURES_DIR, only_path=True),
 	'testssl.output_path': load_fixture('testssl_output', FIXTURES_DIR, only_path=True),
+	'x8.wordlist': 'http_params'
 }
 
 
@@ -142,12 +146,16 @@ def mock_command(cls, inputs=[], opts={}, fixture=None, method=''):
 		fixture = [fixture]
 
 	is_list = isinstance(fixture, list)
+	supports_list = next((loader for loader in cls.item_loaders if getattr(loader, 'list', False)), None)
 	if is_list:
-		for item in fixture:
-			if isinstance(item, dict):
-				mocks.append(json.dumps(item))
-			else:
-				mocks.append(item)
+		if supports_list:
+			mocks.append(json.dumps(fixture))
+		else:
+			for item in fixture:
+				if isinstance(item, dict):
+					mocks.append(json.dumps(item))
+				else:
+					mocks.append(item)
 	else:
 		mocks.append(fixture)