+++ tests/integration/test_tasks.py
@@ -42,7 +42,6 @@ class TestTasks(unittest.TestCase, CommandOutputTester):
 			'ffuf.filter_size': 1987,
 			'feroxbuster.filter_size': 1987,
 			'arjun.wordlist': False,
-			'gau.providers': 'wayback',
 			'h8mail.local_breach': load_fixture('h8mail_breach', INTEGRATION_DIR, only_path=True),
 			'nmap.port': '3000,8080',
 			'nmap.tcp_connect': True,